+++ app/models/meal_plan_recipe.rb
@@ -2,7 +2,8 @@ class MealPlanRecipe < ApplicationRecord
   belongs_to :meal_plan
   belongs_to :recipe
 
-  validates :quantity, numericality: { only_integer: true, greater_than: 0 }
+  validates :quantity,   numericality: { only_integer: true, greater_than: 0 }
+  validates :recipe_id,  uniqueness: { scope: :meal_plan_id }
 
   after_create  :sync_meal_plan_task
   after_destroy :sync_meal_plan_task